Genome to 9-bit and color from genome

+++ Domain/Support/BitSupport.cs
@@ -11,7 +11,7 @@
         /// <param name="b">Input byte data</param>
         /// <param name="bitNumber">Number of bit: 1..8</param>
         /// <returns><see cref="bool"/></returns>
-        public static bool GetBit(this byte b, int bitNumber)
+        public static bool GetBit(this int b, int bitNumber)
         {
             return (b & (1 << bitNumber - 1)) != 0;
         }
